pyproject.toml
[tool.poetry]
name = "qgbif"
version = "0.1.0"
description = ""
authors = ["apierson40 <alexandre.pierson40@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pygbif = "^0.5.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
flake8-isort = "^4.0.0"
black = "^20.8b1"
flake8-builtins = "^1.5.3"
flake8-eradicate = "^1.0.0"
qgis-plugin-ci = "^1.8.4"
pre-commit = "^2.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
